# docs/ahmadkmal-class-6-0.1-swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "defaultTitle",
    "description" : "defaultDescription",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "paths" : {
    "/categories/1" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "put" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "description" : {
                    "type" : "string"
                  },
                  "display_name" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"name\":\"hover\" , \"display_name\":\"hakona\", \"description\":\"for test only\"\n}"
                },
                "1" : {
                  "value" : "{\n     \"description\":\"temon and pomba\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "patch" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "description" : {
                    "type" : "string"
                  },
                  "display_name" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"name\":\"hover\" , \"display_name\":\"hakona\", \"description\":\"for test only\"\n}"
                },
                "1" : {
                  "value" : "{\n     \"description\":\"temon and pomba\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    },
    "/categories" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          },
          "example" : "hover"
        } ],
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "post" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "description" : {
                    "type" : "string"
                  },
                  "display_name" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"name\":\"matata\" , \"display_name\":\"hakona\", \"description\":\"for test only\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    }
  }
}